Add files via upload
Abstract
The primary purpose of fire alarm system is to provide an early warning of fire so that people can be evacuated & immediate action can be taken to stop or eliminate of the fire effect as soon as possible. 
Alarm can be triggered by using detectors or by manual call point (Remotely).
 To alert/evacuate the occupants siren are used. With the Intelligent Building of the rapid development of technology applications, commercial fire alarm market demand growth, the key is to use the bus system intelligent distributed computer system fire alarm system, although installation in the system much easier than in the past , but still cannot meet the modern needs, the installation costs of equipment costs about 33% ~ 70.
 The suggested technique in Fire alarm system used the addressable detectors units besides using the wireless connection between the detector in zones as a slave units and the main control unit as the master unit. 
The system shall include a control panel, alarm initiating devices, notification appliances, and the accessory equipment necessary for a complete functioning fire alarm system. In the wireless fire alarm, individual units are powered by primary & secondary batteries for the communication